[update]セッター処理追加、ヘッダーの修正

+++ Source/EnemyManager.h
@@ -27,13 +27,13 @@ public:
 
 	void Draw();					//描画処理
 
-	void Set_x(int _x);				//x座標をセット
+	void Set_x(float _x);				//x座標をセット
 
-	void Set_y(int _y);				//y座標をセット
+	void Set_y(float _y);				//y座標をセット
 
-	void Set_width(int _width);		//幅をセット
+	void Set_width(float _width);		//幅をセット
 
-	void Set_height(int _height);	//高さをセット
+	void Set_height(float _height);	//高さをセット
 
 	int Get_ActiveFlg(bool);			//アクティブかを受け取る
